chore(footer): removes footer feature flag
