+++ src/Apps/Components/__tests__/AppShell.jest.tsx
@@ -20,7 +20,7 @@ jest.mock("react-tracking", () => ({
 jest.mock("Utils/Hooks/useAuthValidation")
 
 describe("AppShell", () => {
-  it("renders a NavBar", async () => {
+  it("renders a Footer", async () => {
     const { ClientApp } = await buildClientApp({
       history: {
         protocol: "memory",
@@ -43,7 +43,8 @@ describe("AppShell", () => {
         <ClientApp />
       </SystemContextProvider>
     )
-    expect(screen.getByText("Get the Artsy app")).toBeInTheDocument()
+
+    expect(screen.getByText("Meet your new art advisor.")).toBeInTheDocument()
   })
 
   it("calls the matched routes `prepare` function if found", async () => {
