+++ src/Components/Footer/__tests__/Footer.jest.tsx
@@ -1,7 +1,6 @@
 import { MockBoot } from "DevTools/MockBoot"
 import { mount } from "enzyme"
 import { Footer } from "Components/Footer/Footer"
-import { DownloadAppBadge } from "Components/DownloadAppBadges/DownloadAppBadge"
 import { Breakpoint } from "@artsy/palette/dist/themes/types"
 import { useRouter } from "System/Router/useRouter"
 
@@ -23,7 +22,6 @@ describe("Footer", () => {
     it("renders prompts to download the app", () => {
       const wrapper = getWrapper("lg")
       expect(wrapper.find("FooterDownloadAppBanner").length).toEqual(1)
-      expect(wrapper.find(DownloadAppBadge).length).toEqual(2)
     })
 
     it("renders correct routes to partnerships", () => {
@@ -65,7 +63,7 @@ describe("Footer", () => {
 
     it("renders the app download banner", () => {
       const wrapper = getWrapper("lg")
-      expect(wrapper.text()).toContain("Get the Artsy app")
+      expect(wrapper.text()).toContain("Meet your new art advisor.")
     })
 
     it("hides the app download banner if we are on an ignored route", () => {
@@ -74,15 +72,14 @@ describe("Footer", () => {
       }))
 
       const wrapper = getWrapper("lg")
-      expect(wrapper.text()).not.toContain("Get the Artsy app")
+      expect(wrapper.text()).not.toContain("Meet your new art advisor.")
     })
   })
 
   describe("small screen size", () => {
-    it("renders a download banner with a badge and a separate badge", () => {
+    it("renders a download banner", () => {
       const wrapper = getWrapper("xs")
       expect(wrapper.find("FooterDownloadAppBanner").length).toEqual(1)
-      expect(wrapper.find(DownloadAppBadge).length).toEqual(2)
     })
 
     it("renders the CCPA request button", () => {
